+++ CHANGELOG.md
@@ -4,10 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Create `EngagementCardWithImage` component.
 - Fix: Adjust `button` size to follow guidelines on Mobile.
 - Feature: Update `Title` component and add new doc.
 - Feature: Add `RocketIcon` and `RocketCircleIcon` components.
 - Feature: Create `ProgressRing` component.
+- Doc: Add a doc for style `utilities`.
 
 ## [2.64.0] - 2020-04-07